+++ src/components/NewIdea.vue
@@ -80,7 +80,7 @@
 <script lang="ts">
 import { Component, Vue } from 'vue-property-decorator'
 import firebase from 'firebase/app'
-import * as IdeaModel from '@/model/model.ts'
+import { IdeaModel } from '@/model/IdeaModel'
 
 @Component({
   name: 'NewIdea'
@@ -123,7 +123,15 @@ export default class NewIdea extends Vue {
       //   contentInnovationPoint2: this.contentInnovationPoint2,
       //   contentInnovationPoint3: this.contentInnovationPoint3
       // })
-      
+      const item = new IdeaModel()
+      item.title = this.title,
+      item.contentMain= this.contentMain,
+      item.contentSocialPoint= this.contentSocialPoint,
+      item.contentBusinessPoint= this.contentBusinessPoint,
+      item.contentInnovationPoint1= this.contentInnovationPoint1,
+      item.contentInnovationPoint2= this.contentInnovationPoint2,
+      item.contentInnovationPoint3= this.contentInnovationPoint3
+      await item.save()
     } catch (error) {
       console.log('firebase error', error)
     }
